refactor: clipboard 桥脚本文件化——内嵌 BRIDGE 字符串提取为独立 clipboard-bridge.js（review 修订）
内容文件化原则对齐（同 seed 模板/assets）：tapIndex 注入的桥脚本正文从 index.js
内嵌模板字符串搬移至同目录独立文件 clipboard-bridge.js（逐字节零漂移，校验脚本
比对通过），index.js 运行时 readFileSync 读取并包 <script> 注入，只保留包装与
注入逻辑；文件缺失/读取失败降级不注册桥（warn 记录），原生路径不受影响。
cordis 子插件散装分发不经 rspack，文件随包复制，pack terser 按 script 语义压缩。
